+++ src/io/error.rs
@@ -2,8 +2,10 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use rubbl_casatables::CasacoreError;
 use thiserror::Error;
+
+#[cfg(feature = "ms")]
+use rubbl_casatables::CasacoreError;
 
 #[derive(Error, Debug)]
 #[error("bad array shape supplied to argument {argument} of function {function}. expected {expected}, received {received}")]
@@ -16,6 +18,7 @@ pub struct BadArrayShape {
 
 // TODO: there are plenty of panics in ms that need enums
 #[derive(Error, Debug)]
+#[cfg(feature = "ms")]
 pub enum MeasurementSetWriteError {
     /// An error when trying to write to an unexpected row.
     #[error("Tried to write {rows_attempted} rows, but only {rows_remaining} rows are remaining out of {rows_total}")]
@@ -52,12 +55,14 @@ pub enum MeasurementSetWriteError {
     SystemTimeError(#[from] std::time::SystemTimeError),
 }
 
+#[cfg(feature = "ms")]
 impl From<failure::Error> for MeasurementSetWriteError {
     fn from(inner: failure::Error) -> Self {
         Self::RubblError { inner }
     }
 }
 
+#[cfg(feature = "ms")]
 impl From<CasacoreError> for MeasurementSetWriteError {
     fn from(inner: CasacoreError) -> Self {
         Self::CasacoreError { inner }
@@ -65,6 +70,7 @@ impl From<CasacoreError> for MeasurementSetWriteError {
 }
 
 #[derive(Error, Debug)]
+#[cfg(feature = "cfitsio")]
 pub enum UvfitsWriteError {
     /// An error when trying to write to an unexpected row.
     #[error("Tried to write to row number {row_num}, but only {num_rows} rows are expected")]
@@ -90,7 +96,7 @@ pub enum UvfitsWriteError {
 
     /// An error associated with fitsio.
     #[error(transparent)]
-    Fitsio(#[from] mwalib::fitsio::errors::Error),
+    Fitsio(#[from] fitsio::errors::Error),
 
     /// An error when converting a Rust string to a C string.
     #[error(transparent)]
@@ -101,6 +107,7 @@ pub enum UvfitsWriteError {
     StdIo(#[from] std::io::Error),
 }
 
+#[cfg(feature = "cfitsio")]
 impl From<crate::io::uvfits::FitsioOrCStringError> for UvfitsWriteError {
     fn from(e: crate::io::uvfits::FitsioOrCStringError) -> Self {
         match e {
@@ -115,18 +122,22 @@ impl From<crate::io::uvfits::FitsioOrCStringError> for UvfitsWriteError {
 /// All the errors that can occur in file io operations
 pub enum IOError {
     #[error(transparent)]
+    #[cfg(feature = "ms")]
     /// Error derived from [`io::errors::MeasurementSetWriteError`]
     MeasurementSetWriteError(#[from] MeasurementSetWriteError),
 
+    #[cfg(feature = "mwalib")]
     #[error(transparent)]
-    /// Error derived from [`marlu::mwalib::FitsError`]
+    /// Error derived from [`mwalib::FitsError`]
     FitsError(#[from] mwalib::FitsError),
 
     #[error(transparent)]
+    #[cfg(feature = "cfitsio")]
     /// Error derived from [`fitsio::errors::Error`]
-    FitsioError(#[from] mwalib::fitsio::errors::Error),
+    FitsioError(#[from] fitsio::errors::Error),
 
     #[error(transparent)]
+    #[cfg(feature = "cfitsio")]
     /// Error derived from [`io::errors::UvfitsWriteError`]
     UvfitsWriteError(#[from] UvfitsWriteError),
 
@@ -135,9 +146,11 @@ pub enum IOError {
 
     /// From Rubbl
     #[error("Rubbl error {inner:?}")]
+    #[cfg(feature = "ms")]
     RubblError { inner: failure::Error },
 }
 
+#[cfg(feature = "ms")]
 impl From<failure::Error> for IOError {
     fn from(inner: failure::Error) -> Self {
         Self::RubblError { inner }